+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New `issuer list` command used for listing dynamic secrets issuers from Conjur CNJR-9170
 - New `issuer get` command used for retrieving a dynamic secrets issuer from Conjur CNJR-9169
 - New `issuer delete` command used for deleting a dynamic secrets issuer in Conjur CNJR-9168
+- New `issuer update` command used for updating a dynamic secrets issuer in Conjur CNJR-9171
 
 ### Changed
 - Updated Go to 1.24 (CNJR-8641)